Rename and make module closer to template
